requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cu102
torchvision==0.10.0+cu102

numpy==1.19.4
tqdm==4.49.0
matplotlib==3.3.4
braceexpand==0.1.7
einops==0.3.2
imageio==2.9.0
ipython==7.28.0
kornia==0.5.11
omegaconf==2.1.1
Pillow==8.3.2
PyYAML==5.4.1
scikit_learn==1.0
scikit-image==0.18.3
torch_optimizer==0.1.0
torch-tools==0.1.5

#3words
boto3
gitpython

# diffvg
cmake==3.21.3
cssutils==2.3.0
svgpathtools==1.4.2

# can use CompVis/taming-transformers when https://github.com/CompVis/taming-transformers/pull/81 is merged
git+https://github.com/bfirsh/taming-transformers.git@fefe15ac7d0abe69b8d935f1660fdce7567ce11d
git+https://github.com/openai/CLIP@04f4dc2ca1ed0acc9893bd1a3b526a7e02c4bb10
git+https://github.com/pvigier/perlin-numpy@6f077f811f5708e504732c26dee8f2015b95da0c
